backend/pyproject.toml
[project]
name = "netschool-homework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.6"
dependencies = [
    "fastapi>=0.98.0",
    "netschoolapi",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
]

[tool.uv.sources]
netschoolapi = { git = "https://github.com/nota0ua/netschoolapi" }
